Use extension functions instead of OfferUtils and move/rename it to network.bisq.mobile.domain.data.replicated.offer.bisq_easy.BisqEasyOfferVOExtensions
