+++ shared/domain/src/commonMain/kotlin/network/bisq/mobile/domain/data/replicated/offer/DirectionEnumExtensions.kt
@@ -16,9 +16,7 @@
  */
 package network.bisq.mobile.domain.data.replicated.offer
 
-import kotlinx.serialization.Serializable
 
-@Serializable
 object DirectionEnumExtensions {
     val DirectionEnum.isBuy: Boolean
         get() {
